CrewAI/crewai-shopping-cart/pyproject.toml
[project]
name = "crewai-shopping-cart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Arif-Kasim", email = "arif.rozani@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.95.0",
    "google-genai>=1.3.0",
    "google-generativeai>=0.8.4",
    "litellm>=1.61.20",
]

[project.scripts]
kick = "crewai_shopping_cart:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
